Move basic tests
